+++ rs3clans/clans.py
@@ -1,7 +1,6 @@
-# Standard Imports
+from dataclasses import dataclass
 import csv
 
-# Non-standard Imports
 import requests
 
 
@@ -26,6 +25,25 @@ class ClanNotFoundError(Exception):
         return repr(self.value)
 
 
+@dataclass
+class ClanMember:
+    """
+    Represents Member of a Clan
+
+    Args:
+        name (str): Name of the Clan Member
+        rank (str): Rank of the Clan Member
+        exp (int): Clan Exp of the Clan Member
+    """
+    name: str
+    rank: str
+    exp: int
+
+    def __repr__(self) -> str:
+        return f"ClanMember({self.name}, {self.rank}, {self.exp})"
+
+
+@dataclass
 class Clan:
     """Class with attributes of a Runescape Clan using Jagex's RS3 API
 
@@ -33,23 +51,17 @@ class Clan:
 
     ...
 
-    Parameters
-    ----------
-    name : str
-        The name of the clan. Does not need to be case-sensitive.
-    set_exp : bool, optional
-        If True, set.exp will be set with the total Exp from the clan. False by default.
-    set_dict : bool, optional
-        If True, self.member will be set with info from all members from the clan in dict format. True by default.
+    Args:
+        name (str): The name of the clan. Does not need to be case-sensitive.
+    Kwargs:
+        update_stats (bool): If True, updated stats for the Clan will be requested after creating the object and set.
+        set_exp (bool): If True, the `exp` attribute will be set with the total Exp from the clan. False by default.
+        set_dict (bool): If True, `member` attribute (dict) will be set with all members of the clan. True by default.
 
-    Attributes
-    ----------
-    name : str
-        The name of the clan. Set when creating object.
-    exp : int or None
-        The total Exp of the clan, or None if argument set_exp is False (False by default).
-    member : dict or None
-        All info from members in the clan in dictionary format as below::
+    Attributes:
+        name (str): The name of the clan. Set when creating object.
+        exp (int or None): The total Exp of the clan, or None if argument set_exp is False (False by default).
+        member (dict or None): All info from members in the clan in dictionary format as below::
             >>> member_info = {
             ...     'player_1': {
             ...         'exp': 225231234,
@@ -61,79 +73,85 @@ class Clan:
             ...    },
             ... }
         or None if argument set_dict is False (True by default).
-    count : int
-        The number of members in the Clan.
-    avg_exp : int or None
-        The average clan exp per member in the Clan, or None if either set_exp or set_dict arguments are False.
-    hiscores_status_code : int
-        The status code of the GET request to RS3 Clan Hiscores (members lite) API.
+    count (int): The number of members in the Clan.
+    avg_exp (int or None): The average clan exp per member in the Clan.
 
-    Raises
-    ------
-    ClanNotFoundError
-        If an invalid clan name is passed when creating object Clan.
+    Raises:
+        ClanNotFoundError
+            If an invalid clan name is passed when creating object Clan.
     """
 
-    def __init__(self, name: str, set_exp: bool = False, set_dict: bool = True):
-        self.name = name
-        self.exp = None
-        self.member = None
-        self.count = None
-        self.avg_exp = None
+    name: str
+    update_stats: bool = True
+    set_exp: bool = False
+    set_dict: bool = True
 
+    exp: int = None
+    member: dict = None
+    count: int = None
+    avg_exp: float = None
+
+    def __post_init__(self) -> None:
         if not self.name:
             raise TypeError("Clan name cannot be null.")
+        if self.update_stats:
+            self.update()
 
-        self.hiscores_status_code = None
+    def update(self, clan_list: list = None) -> None:
+        """Updates all the Clan stats"""
+        if not clan_list:
+            clan_list = self.parse_clan_list(self.request_clan_list())
 
-        if set_exp:
-            self.exp = self._list_sum()
-        if set_dict:
-            self.member = self.dict_lookup()
+        if self.set_exp:
+            self.exp = self._list_sum(clan_list)
+        if self.set_dict:
+            self.member = self.dict_lookup(clan_list)
             self.count = len(self.member)
-        if set_exp and set_dict:
+        if self.set_exp and self.set_dict:
             self.avg_exp = self.exp / self.count
 
-    def list_lookup(self):
-        """
+    @property
+    def leader(self):
+        """Finds the Clan leader in the clan member dict attribute"""
+        for key, member in self.member.items():
+            if member.rank == 'Owner':
+                return member
 
-        Used for getting all information available from a clan using Rs3's Clan API.
-        Mainly used for calculating the total exp of a clan manually.
+    def request_clan_list(self) -> str:
+        """Requests a Clan list in csv format from RS3's API from the Clan's Name"""
 
-        Returns
-        -------
-        list
-            All players information from a clan in list format like so::
-                >>> clan_list = [
-                ... ['Clanmate', ' Clan Rank', ' Total XP', ' Kills'],
-                ... ['Pedim', 'Owner', '739711654', '2'],
-                ... ['Tusoroxo', 'Deputy Owner', '1132958333', '0'],
-                ... ]
-
-        Raises
-        ------
-        ConnectionError
-            If connecting to the RS3 clan-hiscores/members_lite API was not possible.
-
-        """
         clan_url = f'http://services.runescape.com/m=clan-hiscores/members_lite.ws?clanName={self.name}'
 
         with requests.Session() as session:
             download = session.get(clan_url)
-            decoded = download.content.decode('windows-1252')
-            self.hiscores_status_code = download.status_code
-            if download.status_code != 200:
-                raise ConnectionError(
-                    f'Not able to connect to RS3 clan-hiscores/members_lite API. Status code: {download.status_code}'
-                )
-            clan_list = list(csv.reader(decoded.splitlines(), delimiter=','))
-            if clan_list[0][0] != "Clanmate":
-                raise ClanNotFoundError(f"Couldn't find clan: {self.name}")
-            for row in clan_list:
-                row[0] = row[0].replace(r"\xa0", " ")
-            return clan_list
+            return download.content.decode('windows-1252')
 
-    def dict_lookup(self, rank_key: str = "rank", exp_key: str = "exp"):
+    def parse_clan_list(self, decoded: str) -> list:
+        """
+        Parses a Clan List csv and returns it in a Python List
+
+        Returns:
+            list. All players information from a clan in list format like so::
+                    >>> clan_list = [
+                    ... ['Clanmate', ' Clan Rank', ' Total XP', ' Kills'],
+                    ... ['Pedim', 'Owner', '739711654', '2'],
+                    ... ['Tusoroxo', 'Deputy Owner', '1132958333', '0'],
+                    ... ]
+        """
+
+        clan_list = list(csv.reader(decoded.splitlines(), delimiter=','))
+
+        if clan_list[0][0] != "Clanmate":
+            raise ClanNotFoundError(f"Couldn't find clan: {self.name}")
+
+        for row in clan_list:
+            # Replacing non-breaking spaces with actual spaces
+            print(row[0])
+            row[0] = row[0].replace(r"\xa0", " ")
+        return clan_list
+
+    @staticmethod
+    def dict_lookup(clan_list: list) -> dict:
         """
 
         Used for getting all information available from a clan using Rs3's Clan API.
@@ -141,109 +159,63 @@ class Clan:
         Contrary to `list_lookup` it returns it as a Dictionary format instead.
         The dict format makes easier to find info specific to certain members of the Clan instead of looping over it.
 
-        Parameters
-        ----------
-        rank_key : str
-            The key for Clan Rank to be used in the returned dictionary with all members info of clan.
-        exp_key : str
-            The key for Clan Exp to be used in the returned dictionary with all members info of clan.
+        Args:
+            clan_list (list): The clan list to make the dictionary for
 
-        Returns
-        --------
-        dict
-            Information from all members in the clan in dictionary format for ease of access.
+        Returns:
+            dict[str, ClanMember]. Information from all members in the clan in dictionary format for ease of access.
         """
-        clan_list = self.list_lookup()
-        clan_dict = {}
-        for row in clan_list[1:]:
-            user_rank = row[1]
-            username = row[0].replace('\xa0', ' ')
-            user_exp = int(row[2])
+        return {row[0]: ClanMember(row[0], row[1], int(row[2])) for row in clan_list[1:]}
 
-            clan_dict[username] = {
-                rank_key: user_rank,
-                exp_key: user_exp
-            }
-        return clan_dict
-
-    def get_member(self, name: str):
+    def get_member(self, name: str) -> ClanMember:
         """
 
         Used for searching information about a clan member by passing in its name case insensitively.
 
-        Parameters
-        ----------
-        name : str
-            The name of the player to be searched in clan.
+        Args:
+            name (str): The name of the player to be searched in clan.
 
-        Returns
-        --------
-        dict
-            Information of the member passed in, with the keys ['exp'] and ['rank'] (or any custom keys you might have
-            passed when creating Clan object.
+        Returns:
+            ClanMember. ClanMember object of the found (if found) player.
         """
         for key, value in self.member.items():
             if key.lower() == name.lower():
-                value['name'] = key
                 return value
 
-    def _list_sum(self):
+    @staticmethod
+    def _list_sum(clan_list: list) -> int:
         """
-        Gets a clan list from `list_lookup` and sums the total exp of the clan.
 
-        Returns
-        -------
-        int
-            The total Exp of the Clan.
+        Returns the Exp sum of a clan list.
+
+        Returns:
+            int. The total Exp of the Clan.
         """
-        clan_list = self.list_lookup()
         return sum(int(row[2]) for row in clan_list[1:])
 
-    def _dict_sum(self):
-        """
-        Gets a clan dictionary from :func: `dict_lookup` and sums the total exp of the clan.
-
-        .. deprecated:: 0.4.0
-            Use `list_sum` as it's much faster.
-
-        Returns
-        -------
-        int
-            The total Exp of the Clan.
-        """
-        clan_dict = self.dict_lookup()
-        return sum(members['exp'] for members in clan_dict.values())
-
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"Clan(name={repr(self.name)}, exp={self.exp}, avg_exp={self.avg_exp}, count={self.count})"
 
-    def __iter__(self):
+    def __iter__(self) -> ClanMember:
         """
+
         Iterates through `self.member` dict attribute, if `set_dict` was passed True when instancing `Clan` object.
 
-        Yelds
-        -----
-        tuple
-            Tuple contains the key of each item in `Clan.member` attribute and then its values::
+        Yields:
+            ClanMember. ClanMember object of each member in the clan::
                 >>> import rs3clans
                 >>> clan = rs3clans.Clan('atlantis')
                 >>> for member in clan:
-                ...     print(member)
+                ...     print(f"{member} - {member.name}")
                 ...
-                ('Pedim', {'rank': 'Owner', 'exp': 962627411})
-                ('Tusoroxo', {'rank': 'Deputy Owner', 'exp': 1157686923})
-                ('Acriano', {'rank': 'Overseer', 'exp': 1857267476})
-                ('Cogu', {'rank': 'Overseer', 'exp': 1576814120})
-                ('Black bullet', {'rank': 'Overseer', 'exp': 909580894})
-                ('NRiver', {'rank': 'Overseer', 'exp': 1043065027})
+                ClanMember(Pedim, Owner, 1249520826) - Pedim
+                ClanMember(Acriano, Overseer, 1903276564) - Acriano
+                ClanMember(Cogu, Overseer, 1829449412) - Cogu
+                ClanMember(Black bullet, Overseer, 1100767386) - Black Bullet
+                ClanMember(NRiver, Overseer, 1090093362) - NRiver
+                ClanMember(Kurenaii, Overseer, 395850997) - Kurenaii
                 (...)
-                >>> clan = rs3clans.Clan('atlantis', set_dict=False)
-                >>> for member in clan:
-                ...     print(member)
-                >>>
-                ...
         """
         if self.member:
             for key, value in self.member.items():
-                yield key, value
-        return
+                yield value