+++ rs3clans/players.py
@@ -1,10 +1,10 @@
-# Standard Imports
+from dataclasses import dataclass
 import json
 
-# Non-standard Imports
 import requests
 
 
+@dataclass
 class Player:
     """Class with attributes of a Runescape Player using Jagex's RS3 API
 
@@ -12,68 +12,62 @@ class Player:
 
     ...
 
-    Parameters
-    ----------
-    name : str
-        Name of the Player, gets set to case-sensitive name later if his RuneMetrics profile is not private.
+    Args:
+        name (str): Name of the Player, gets set to case-sensitive name later if his RuneMetrics profile is not private.
         If you're also working with clans and the player's profile is private, name should be passed case-sensitively.
         Otherwise some clan attributes will not be available.
-    runemetrics : bool
-        Where or not Runemetrics player info should be set or not, slows down creating the object, but provides more
-        attributes and methods.
 
-    Attributes
-    ----------
-    name : str
-        Name of the Player, gets set to case-sensitive name later if his RuneMetrics profile is not private.
-    metrics_info : dict
-        Runemetrics Info in dictionary format about the player.
-    exp : int or None
-        Total Exp of the player, or None if private_profile is True.
-    combat_level : int or None
-        Total Combat Level of the player, or None if private_profile is True.
-    total_level : int or None
-        Total Skill Level of the player, or None if private_profile is True.
-    quests_not_started : int
-        Number of quests the player has not started.
-    quests_started : int
-        Number of the quests the player has started.
-    quests_complete : int
-        Number of quests the player has completed.
-    private_profile : bool
-        Whether the player has a private RuneMetrics profile or not.
-    exists : bool
-        Whether the player exists or not.
-    details_status_code : int
-        The status code of the GET request to RS3 Player Details API
-    runemetrics_status_code : int
-        The status code of the GET request to RS3 Player Runemetrics API
+        runemetrics (bool): Whether or not Runemetrics player info should be set or not, slows down creating the object,
+        but provides more attributes and methods.
 
-    Raises
-    ------
-    ConnectionError
-        If connecting to any of RS3's API related to players fails.
+    Attributes:
+        name (str): Name of the Player, gets set to case-sensitive name later if his RuneMetrics profile is not private.
+
+        metrics_info (str): Runemetrics Info in dictionary format about the player.
+
+        exp (int or None): Total Exp of the player, or None if private_profile is True.
+
+        combat_level (int or None): Total Combat Level of the player, or None if private_profile is True.
+
+        total_level (int or None): Total Skill Level of the player, or None if private_profile is True.
+
+        quests_not_started (int): Number of quests the player has not started.
+
+        quests_started (int): Number of the quests the player has started.
+
+        quests_complete (int): Number of quests the player has completed.
+
+        private_profile (bool): Whether the player has a private RuneMetrics profile or not.
+
+        exists (bool): Whether the player exists or not.
+
+        details_status_code (int): The status code of the GET request to RS3 Player Details API
+
+        runemetrics_status_code (int): The status code of the GET request to RS3 Player Runemetrics API
+
+    Raises:
+        ConnectionError
+            If connecting to any of RS3's API related to players fails.
     """
+    name: str
+    runemetrics: bool = True
 
-    def __init__(self, name: str, runemetrics: bool = True):
+    metrics_info: dict = None
+    exp: int = None
+    combat_level: int = None
+    total_level: int = None
+    quests_not_started: int = None
+    quests_started: int = None
+    quests_complete: int = None
+    skill_values: list = None
+    details_status_code: int = None
+    runemetrics_status_code: int = None
+    private_profile: bool = True
+    exists: bool = True
 
-        self.name = name
-        self.metrics_info = None
-        self.exp = None
-        self.combat_level = None
-        self.total_level = None
-        self.quests_not_started = None
-        self.quests_started = None
-        self.quests_complete = None
-        self.skill_values = None
-        self.private_profile = True
-        self.exists = True
-
+    def __post_init__(self) -> None:
         if not self.name:
             raise TypeError("Player name cannot be null.")
-
-        self.details_status_code = None
-        self.runemetrics_status_code = None
 
         self.info = self._dict_info()
         self.suffix = self.info['is_suffix']
@@ -84,7 +78,7 @@ class Player:
         if not self._valid_username(self.name):
             self.exists = False
 
-        if runemetrics:
+        if self.runemetrics:
             self.set_runemetrics_info()
 
         try:
@@ -93,20 +87,16 @@ class Player:
             self.clan = None
 
     @staticmethod
-    def _valid_username(username: str):
+    def _valid_username(username: str) -> bool:
         """
         Checks if a username is in the valid character limit set by Jagex
         and if it doesn't have any invalid characters in it.
 
-        Parameters
-        ----------
-        username : str
-            The username to be checked if valid or not
+        Args:
+            username (str): The username to be checked if valid or not
 
-        Returns
-        -------
-        bool
-            True if the username is valid, False if it is not
+        Returns:
+            bool. True if the username is valid, False if it is not
         """
         if len(username) > 12:
             return False
@@ -119,22 +109,20 @@ class Player:
             return False
         return True
 
-    def _raw_info(self):
+    def _raw_info(self) -> str:
         """
         Gets JsonP information on a player, in string format.
 
-        Returns
-        -------
-        str
-            JsonP information on player, converted to string.
+        Returns:
+            str. JsonP information on player, converted to string.
 
-        Raises
-        ------
-        ConnectionError
-            If connecting to the RS3 Runemetrics profile API was not possible.
+        Raises:
+            ConnectionError
+                If connecting to the RS3 Runemetrics profile API was not possible.
 
         ------
-        .. note:: Has to be formatted correctly into Json or other formats to be worked with properly first.
+        .. note::
+            Has to be formatted correctly into Json or other formats to be worked with properly first.
         """
 
         info_url = (f"http://services.runescape.com/m=website-data/"
@@ -145,15 +133,13 @@ class Player:
             raise ConnectionError(f'Not able to connect to RS3 playerDetails API. Status code: {response.status_code}')
         return str(response.content)
 
-    def _dict_info(self):
+    def _dict_info(self) -> dict:
         """
         Gets the raw string info from self._raw_info() and formats it into Dictionary format.:
         Used to set self.info.
 
-        Returns
-        -------
-        dict
-            Dictionary is formatted as follows::
+        Returns:
+            dict. Dictionary is formatted as follows::
                 >>> player_info = {
                 ... 'is_suffix': True,
                 ... 'recruiting': True,
@@ -161,16 +147,15 @@ class Player:
                 ... 'clan': 'Atlantis',
                 ... 'title': 'the Liberator'
                 ... }
-            is_suffix : bool
-                If the player's title is a Suffix or not
-            recruiting : bool
-                If the player's clan is set as Recruiting or not
-            name : str
-                The player's name, passed as is when creating object Player
-            clan : str
-                The player's clan name
-            title : str
-                The player's current title
+            is_suffix (bool): If the player's title is a Suffix or not
+
+            recruiting (bool): If the player's clan is set as Recruiting or not
+
+            name (str): The player's name, passed as is when creating object Player
+
+            clan (str): The player's clan name
+
+            title (str): The player's current title
         """
         str_info = self._raw_info()
         info_list = []
@@ -186,20 +171,18 @@ class Player:
         del info_dict['isSuffix']
         return info_dict
 
-    def set_runemetrics_info(self):
+    def set_runemetrics_info(self) -> None:
         """
         Sets Runemetrics info attributes for the Player.
 
-        Attributes set
-        --------------
-        name : str
-            Case-sensitive player username.
-        exp : int
-            Total Exp of the Player.
-        combat_level : int
-            Total Combat Level of the Player.
-        total_level : int
-            Total Skill Level of the Player.
+        Attributes set:
+            name (str): Case-sensitive player username.
+
+            exp (int): Total Exp of the Player.
+
+            combat_level (int): Total Combat Level of the Player.
+
+            total_level (int): Total Skill Level of the Player.
         """
         # If user's runemetrics profile is private, self.name will be the same as passed when creating object.
         # Otherwise it will get the correct case-sensitive name from his runemetrics profile.
@@ -261,15 +244,12 @@ class Player:
             ... "loggedIn":"false"
             ... }
 
-        Returns
-        ------
-        dict or bool
-            Runemetrics player info if the player profile is public, or False if it is Private.
+        Returns:
+            dict or bool. Runemetrics player info if the player profile is public, or False if it is Private.
 
-        Raises
-        ------
-        ConnectionError
-            If connecting to the RS3 Runemetrics profile API was not possible.
+        Raises:
+            ConnectionError
+                If connecting to the RS3 Runemetrics profile API was not possible.
         """
         user_name = self.name.replace(' ', '%20')
         info_url = f"https://apps.runescape.com/runemetrics/profile/profile?user={user_name}&activities=0"
@@ -306,21 +286,20 @@ class Player:
             >>> player.skill('AtTaCk').rank
             68311
 
-        Parameters
-        ----------
-        skill : str or int
-            Skill name or id to get info of
+        Args:
+            skill (str or int): Skill name or id to get info of
 
-        Returns
-        ----------
-        Skill object
+        Returns:
+            Skill.
             Attributes of Skill
             --------------------
-                exp (or 'xp') : Total exp the Player has in that skill
+                exp (int): Total exp the Player has in that skill
 
-                level : The level the Player has in that skill
+                xp (int): Total exp the Player has in that skill
 
-                rank : The rank the Player has in that skill
+                level (int): The level the Player has in that skill
+
+                rank (int): The Rank the Player has in that skill
         """
         skills_index = {
             0: 'attack',
@@ -370,26 +349,43 @@ class Player:
         return f"Player(name={repr(self.name)}, clan={repr(self.clan)}, exists={self.exists})"
 
 
+@dataclass
 class Skill:
-    def __init__(self, name: str, skill: dict):
-        self.name = name
-        try:
-            self.exp = skill['xp'] / 10
-        except KeyError:
-            self.exp = None
+    """
+    Represents a RuneScape 3 Skill
+
+    Args:
+        name (str): Name of the Skill
+
+        skill (dict): Skill in dictionary format as it is in RuneScape's 3 API
+
+    Attributes:
+        exp (int): Total Exp the player has in the skill
+
+        xp (int): Total Exp the player has in the skill
+
+        level (int): Total Level the player has in the skill
+
+        rank (int): Rank the player has in the skill
+
+        id (int): The ID of the Skill
+    """
+    name: str
+    skill: dict
+
+    exp: int = None
+    xp: int = None
+    level: int = None
+    rank: int = None
+    id: int = None
+
+    def __post_init__(self):
+        self.exp = self.skill.get('xp') / 10
         self.xp = self.exp
-        try:
-            self.level = skill['level']
-        except KeyError:
-            self.level = None
-        try:
-            self.rank = skill['rank']
-        except KeyError:
-            self.rank = None
-        try:
-            self.id = skill['id']
-        except KeyError:
-            self.id = None
+
+        self.level = self.skill.get('level')
+        self.rank = self.skill.get('rank')
+        self.id = self.skill.get('id')
 
     def __str__(self):
         return self.name